requirements.txt
flask
flask-cors
transformers
torch
nltk
pdfplumber
python-pptx
gunicorn
